+++ NEWS.md
@@ -2,6 +2,9 @@
 * Improved method for calculating the object center of gravity for automated
 ROI placement.
 * Added method to calcuate the maximum background (MBG) signal.
+* Added an option (pix_dim) to override the image voxel dimensions.
+* Mean background image added to output - TODO.
+* Background TC added to output - TODO.
 
 # fmriqa 0.2.0
 * Added an argument for the polynomial detrending order and changed the default